lambda-s3-gemini-slack/requirements.txt
boto3
requests
pytz
google-generativeai
